feat(home): T4.13 — move proxy groups into pushed subview
Home now shows a single "Proxy Groups" row with the count + chevron;
tapping pushes the new ProxyGroupsView which owns the per-group picker
(ProxyGroupCard, DelayBadge, fetch/select/ping). Row is disabled and
shows "—" when count is zero (not connected, no profile, fetch error).

HomeView keeps the stage-keyed + replayGeneration-keyed refresh tasks
but only stores groupCount: Int. Per-group expansion, ping inflight set,
and the inline placeholder card all move to the subview. Existing
accessibility identifiers (home.group.<slug>, home.proxy.<g>.<p>) are
preserved inside the subview; new IDs added: home.nav.proxyGroups,
home.proxyGroups.count.
